+++ src/components/App.css
@@ -1,0 +1,3 @@
+h1 {
+  font-family: 'Megrim', cursive;
+}
