+++ src/components/App.jsx
@@ -1,11 +1,14 @@
 import React from 'react';
 import ImageSlider from './ImageSlider';
 import './App.css';
-import data from './data';
+import data from '../data/images';
 
 function App() {
   return (
-    <ImageSlider images={data}/>
+    <>
+      <h1>Let it snow, let it snow</h1>
+      <ImageSlider images={data}/>
+    </>
   );
 }
