+++ src/index.css
@@ -1,21 +1,1 @@
 @import url('https://fonts.googleapis.com/css?family=Megrim&display=swap');
-
-h1 {
-  font-family: 'Megrim', cursive;
-}
-.slider {
-  display: flex;
-  flex-flow: row wrap;
-  width: 380px;
-  justify-content: space-between;
-}
-img.slider_arrow{
-  display: block;
-  align-self: center;
-}
-ul.slider_list {
-  padding: 0px;
-}
-.slider_list_item {
-  list-style: none;
-}
